refactor: move content-v2 apis
