+++ lambdas/src/apis/content-v2/controllers/translator.ts
