+++ lambdas/src/controllers/routes.ts
@@ -9,6 +9,7 @@ import {
 import { getEmotesByOwnerHandler, getEmotesHandler } from '../apis/collections/controllers/emotes'
 import { getWearablesByOwnerHandler, getWearablesHandler } from '../apis/collections/controllers/wearables'
 import { OffChainWearablesManager } from '../apis/collections/off-chain/OffChainWearablesManager'
+import { getContents, getInfo, getScenes } from '../apis/content-v2/controllers/translator'
 import { Bean, Environment, EnvironmentConfig } from '../Environment'
 import { createFetchComponent } from '../ports/fetcher'
 import { createThirdPartyAssetFetcher, ThirdPartyAssetFetcher } from '../ports/third-party/third-party-fetcher'
@@ -16,7 +17,6 @@ import { DAOCache } from '../service/dao/DAOCache'
 import { SmartContentClient } from '../utils/SmartContentClient'
 import { SmartContentServerFetcher } from '../utils/SmartContentServerFetcher'
 import { TheGraphClient } from '../utils/TheGraphClient'
-import { getContents, getInfo, getScenes } from './handlers/content-v2/handlers'
 import { getCatalystServersList, getDenylistedNamesList, getPOIsList } from './handlers/contracts/handlers'
 import { validateSignature } from './handlers/crypto/handlers'
 import { hotScenes, realmsStatus } from './handlers/explorer/handlers'
